2019/10/31 上午日常开发
商品添加开发

+++ newtcl/src/cn/newtcl/service/impl/ImTypeService.java
@@ -101,6 +101,11 @@ public class ImTypeService implements TypeService {
 			re.setMessage("未给出主键 id");
 			return re ;
 		}
+		if(type.getId() <= 4){
+			re.setCode("-1");
+			re.setMessage("受保护数据不能删除!");
+			return re ;
+		}
 		int temp = typeMapper.delete(type);
 		if(temp > 0){
 			re.setCode("1");
